Update AGRIF sub-grid coordinates re: nesting tools bug.
From email sent by Micahel:

Christoph is spot on (below) about the bug in the agrif_create_coordinates program, it writes the wrong coordinates for u and v points. It seems remarkable that this has not been noticed before (including by myself.....) but after poking around a but, I think I understand why. First, the nesting tools don't use the u and v points to compute the scale factors, rather they interpolate the parent grid scale factors in a fashion akin to our unstagger function in salishsea tools (that is, it does not use any absolute coordinates), and then it divides them by the refinement factor -- so the scale factors were calculated right. Second, since we're using jphgr_msh=0 in NEMO, the model loads the scale factors in rather than computing them, and also it makes no use of the u,v coordinates during the time stepping calculations. Naturally, since the t points were not wrong, the T&S interpolation and bathy were unaffected.

So on one hand, it's a bug that doesn't affect the NEMO simulation when you ramp up a simulation from rest. On the other hand, there are at least two cases where it might matter. The first is when you create a child-grid restart file with agrif_create_restart -- there should be a small error in the velocities because they'd be interpolated to slightly wrong locations. Second, if you're relying on the u and v points in the coordinates file for visualization purposes, or for any other diagnostic calculation of fluxes etc, the results will be slightly off.

This will affect the Baynes Sound / Haro Strait subdomains, so please regenerate the coordinates file(s) with the fix here [1], and also regenerate the child-grid restart files. That said, the wrong u,v should be off by less than one fine grid cell, so the error in interpolation should be rather small, but not zero.

Bug was fixed by https://bitbucket.org/mdunphy/nestingtools/commits/06dd49ed2