Add coordinates for 202405 bathymetry
This is the 201702 coordinates with 10 rows of grid cells added on
the south (y=0) side. That was done to include the waters of
south Puget Sound in the Nisqually River estuary region that
have been excluded since the original seagrid-derived
coordinates.

Please see
https://github.com/SalishSeaCast/analysis-doug/blob/main/notebooks/2xrez-202111/coordinates.ipynb
for details of the calculation of this coordinates file.